更新login流程

+++ Modules/Base/Contract/Database/BaseMigration.php
@@ -5,9 +5,15 @@ namespace Modules\Base\Contract\Database;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Support\Facades\Schema;
 use Modules\Base\Constants\ConnectionConfigConstants;
+use Modules\Base\Exception\BaseException;
+use Modules\Error\Constants\ErrorCode;
 
 abstract class BaseMigration extends Migration
 {
+    const MODE_CREATE = 'create';
+
+    const MODE_UPDATE = 'update';
+
     /** @var string  */
     protected $table = '';
 
@@ -17,11 +23,16 @@ abstract class BaseMigration extends Migration
     /** @var string  */
     protected $connection = ConnectionConfigConstants::MAIN_CONNECTION_NAME;
 
+    /** @var  string */
+    protected $mode = 'create';
+
     public function up()
     {
-        Schema::connection($this->getConnection())->create($this->table, $this->tableSchema());
-        \DB::connection($this->getConnection())
-            ->statement("ALTER TABLE `{$this->table}` comment '{$this->tableComment}'");
+        Schema::connection($this->getConnection())->{$this->getModeFunc()}($this->table, $this->tableSchema());
+        if ($this->mode == self::MODE_CREATE) {
+            \DB::connection($this->getConnection())
+                ->statement("ALTER TABLE `{$this->table}` comment '{$this->tableComment}'");
+        }
         $this->run();
     }
 
@@ -38,8 +49,25 @@ abstract class BaseMigration extends Migration
      */
     public function down()
     {
-        \DB::connection($this->getConnection())->statement('SET FOREIGN_KEY_CHECKS = 0');
-        Schema::connection($this->getConnection())->dropIfExists($this->table);
-        \DB::connection($this->getConnection())->statement('SET FOREIGN_KEY_CHECKS = 1');
+        if ($this->mode == self::MODE_CREATE) {
+            \DB::connection($this->getConnection())->statement('SET FOREIGN_KEY_CHECKS = 0');
+            Schema::connection($this->getConnection())->dropIfExists($this->table);
+            \DB::connection($this->getConnection())->statement('SET FOREIGN_KEY_CHECKS = 1');
+        }
+    }
+
+    private function getModeFunc()
+    {
+        switch ($this->mode) {
+            case self::MODE_CREATE:
+                return 'create';
+            case self::MODE_UPDATE:
+                return 'table';
+            default:
+                throw new BaseException(
+                    'base migrate mode error',
+                    ErrorCode::BASE_MIGRATE_MODE_ERROR
+                );
+        }
     }
 }
